Fixed with JPA

+++ src/main/java/co/pragra/learning/librarymgment/rest/BookApi.java
@@ -1,7 +1,7 @@
 package co.pragra.learning.librarymgment.rest;
 
 import co.pragra.learning.librarymgment.entity.Book;
-import co.pragra.learning.librarymgment.repo.BookRepo;
+import co.pragra.learning.librarymgment.repo.BookRepository;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.Collections;
@@ -11,43 +11,45 @@ import java.util.Optional;
 @RestController
 public class BookApi {
 
-    private BookRepo repo;
+    private BookRepository repo;
 
-    public BookApi(BookRepo repo) {
+    public BookApi(BookRepository repo) {
         this.repo = repo;
     }
 
     @PostMapping("/api/book")
     public Book createBook(@RequestBody Book book) {
-        if(book.getId()==0) {
-            book.setId(((int) Math.floor(Math.random() * 10000)));
-        }
-        return  repo.create(book);
+        return  repo.save(book);
     }
 
     @GetMapping("/api/book")
     public List<Book> getAll() {
-        return repo.getAll(Collections.emptyMap());
+        return repo.findAll();
     }
 
     @GetMapping("/api/book/id/{id}")
     public Optional<Book> getById(@PathVariable int id) {
-        return repo.getById(id);
+        return repo.findById(id);
     }
 
     @GetMapping("/api/book/isbn/{isbn}")
     public Optional<Book> getByIsbn(@PathVariable String isbn) {
-        return repo.getByIsbn(isbn);
+        return repo.findByIsbn(isbn);
     }
 
     @DeleteMapping("/api/book/id/{id}")
     public Book deleteById(@PathVariable int id) {
-        return repo.deleteById(id);
+        Optional<Book> byId = repo.findById(id);
+        if(byId.isPresent()) {
+            repo.deleteById(id);
+        }
+
+      return  byId.get();
     }
 
     @PutMapping("/api/book")
     public Book update(@RequestBody Book book){
-        return repo.update(book);
+        return repo.save(book);
     }
 
 }
